add multi-image support to mvpnet
Add multi-image support to mvpnet instead of limiting guests
to just the boot drive.   This changes the CLI as follows:

 - allow multiple "-i" image flags on the command line.  Each instance
   adds a new storage drive to the guest.

 - remove flags -I and -R -- use per-image "mvpctl=" property instead.

 - the "-i" flag supports optional MPI rank filtering

The "-i" flag now takes a list of comma separated 'property=value'
pairs.  The arg format for "-i" is one of:

  -i /path/to/image.img,property1=value1,property2=value2,...

  or

  -i property1=value1,property2=value2,...

The first format (with the path to an image file) is used when
mvpnet is directly managing the image filename and location.
In this case, mvpnet defines it's own optional "mvpctl=" property
that allows additional image handling to be specified on the command
line.

The second format is passed down directly to qemu's "-drive" flag.
It should have a "file=" property and should NOT have a "mvpctl="
property.

For the first format, the "mvpctl=" property replaces the old -I
and -R flags.  The characters in mvpctl have the following meanings:

   c - copy image to rundir (-r must be set)
   j - add job to filename in rundir  (use with 'c')
   r - add rank to filename in rundir (use with 'c')
   d - remove image from rundir at exit (use with 'c')
   J - add job to image filename (-i)
   R - add rank to image filename (-i)
   D - remove image (-i) at exit
   p - write-protect image file (read-only=on)
   s - open file in snapshot mode (snapshot=on)

relevant code changes:
 - getopt processing changed to match above
 - we accumulate -i args in mopt.image (now a strvec)
 - -i processing moved to qemucli.c
 - added qemucli_procarg() to parse/process qemu-style prop=val args
 - added qemucli_storage_img() to handle mvpctl for an image
 - qemucli_storage_cfg() walks the list of -i args and processes them
 - copyfile() now allows you to specify additional open(2) flag values