cli change: change console -c to -C, add -c for cloudinit
Add shortcut flag when we configure with cloud-init:

   -c [val]

where val is an image file or a directory with cloud-init config info.
these get translated into qemu "-drive" flags.  for the directory case,
we add it as a fat filesystem with the appropriate "file.label=cidata" info.
The cloudinit -drive arg is always inserted as the second -drive arg to qemu
(so you can specify "-c" to mvpnet before "-i" and it will get inserted
in the correct place).