do second loan rather than memcpy 2nd bcast frame in proc qframe
When the guest starts a broadcast (fdio_process_qframe() case 3)
and we need to forward the frame to more than one child in the
broadcast tree, just add a second loan to the fbuf rather than
memcpying it to new space allocated in fbm_bcast via
fbufmgr_loan_newframe().   there is no need to copy the frame here.