fdio_read_stdin: only relay data if qstdin is open
have fdio_read_stdin() only relay data if qstdin is open (i.e. >= 0).
allows for the case where we have closed qstdin to send an EOF to qemu
and we are waiting for it respond by exiting.  in this case we just
drop the input data since we've already released qstdin to send the EOF.