add MPI drain shutdown code
to support shutdowns started by mvpnet, fdio_thread needs to be able to
queue a set of shutdown ethernet frames for sending (in the MPI sendq),
disable queuing of any additional frames to send, and wait for the MPI
thread to complete its async sending of pending frames.  once this is
done, we can send an EOF to qemu's stdin (typically monwrap) to have
it shutdown.   once qemu shuts down, mvpnet can shutdown.

add low-level mpi send queue drain code to support this.  data structure
and define changes:

 - add mpisq_draindown to mvp_queuing: set non-zero if we are in draindown

 - add mpisndequeued to mvp_queuing: tracks the number of send queue entries
   the MPI thread has dequeued from mpisendq and is currently sending with
   MPI_Isend().

 - add FDIO_NOTE_DRAINED notification: sent when in draindown mode and
   the last send queue entry (sqe) being used is released.  at this point
   we send EOF to qemu by closing its stdin.

API:
to trigger the draindown mode, call:

    int mvp_sq_draindown(struct mvp_queuing *mq)

this will set mpisq_draindown and return the number of sends we need
to complete before we can send EOF to qemu.   if it returns 0, then
we can continue now.  otherwise we must wait for the FDIO_NOTE_DRAINED
notification before advancing.

so the sqe life cycle is:

  - mvp_sq_queue() - fdio_thread allocates sqe, fills it out, puts on sendq,
                     does mpisqlen++, will fail if sqe alloc fails or
                     if we are in draindown mode now (in draindown we
                     expect it to fail, since sends are now blocked).

  - mvp_sq_dequeue() - mpi thread dequeues sqe, starts async MPI_Isend
                       does mpisqlen--, mpisndequeued++

  - mvp_sq_release() - mpi thread finishes MPI_Isend(), frees sqe
                       does mpisndequeued--
                       sends FDIO_NOTE_DRAINED notification to fdio if
                       we are in draindown and just zeroed the counters.