add optional user callback to fdforkprog(): after fork, before exec
Add an optional user callback to fdforkprog() after the fork and
before the exec.   This allows cleanup in the child before we exec
the new program.   If there is an error, the callback can exit
(thus preventing the exec).   A mvp_closelog() callback that we
can use with this hook to close the mlog file descriptors before
we exec (since these file descriptors are not created by mvpnet code
we cannot count on them being marked close-on-exec).