ignore EOF on stdin by default, add -E to look for it
the handling of a rank's stdin is MPI implementation dependent.
we were using EOF on stdin to signal a shutdown, but there are
cases where MPI starts us with stdin closed so we cannot count
on this.   update code to start ignoring stdin if it gets an
EOF on it.  the old behavior can be re-enabled by using the new
-E flag.

some data from mpich and openmpi from a test program called "stdin":
  ./stdin                          # stdin=/dev/pts/0
  mpirun -n 1 ./stdin              # stdin=a pipe
  mpirun -n 1 ./stdin < /dev/null  # stdin=a pipe,read side closed
                                   #  so read is always EOF
  mpirun -n 3 ./stdin
          # mpich: each rank has an open pipe as stdin
          # openmpi: rank 0 has open pipe as stdin
          #          ranks 1,2 have /dev/null as stdin

  mpirun -n 3 ./stdin < /dev/null
          # mpich: rank 0 has pipe with read size closed as stdin
          #        ranks 1,2 have open pipe as stdin
          # openmpi: rank 0 has pipe with read size closed as stdin
          #          ranks 1,2 have /dev/null as stdin

  mpich: mpirun does not have --stdin flag
  openmpi: has --stdin flag: allows you to move stdin to
           a specific rank, all the ranks, or none of the ranks

we were testing with mpich, and that worked fine until we tried
to run a test with input from /dev/null and discovered that rank 0
was getting EOF and exiting early.