add tfinishpid()
tfinishpid() waits for an exiting pid to finish its exit and
return a wait status.  the caller should have already asked
the pid to exit (e.g. by sending it an EOF by closing its stdin
or by sending it a signal).   if the pid does not exit in the
number of seconds specified we will send it a SIGKILL to ensure
it is no longer running.  we avoid using ITIMER_REAL/SIGALRM so
this should be ok in a threaded environment (e.g. vs twaitpid()).
return 0 on a successful wait, otherwise -1.